requirements.txt
asgiref==3.4.1
Django==3.2.7
django-debug-toolbar==3.2.2
djangorestframework==3.12.4
psycopg2==2.9.1
pytz==2021.1
sqlparse==0.4.2
